Began major restructuring
Started to shift the bot to be structured like a python package that way it can be installed and started as one. This goes in hand with some restructuring of the directories to make it more clear what various files are actually for. This still needs to be tested as there is no guaranteed that this actually runs at all since the refactor given there were major structural changes. Relative imports and other things may be broken as well.