New config location
The bot's configuration class now exists within the bot package. The initialization method has been cleaned up to reduce issues. Errors will now be thrown when it makes sense and warnings for missing envars. The config file is now global from the bot package and doesn't need to be accessed from the bot instance, but can still be done for compatibility.