pagination demo
Started to come up with a system for pagination of data from an sqlite database. The method I have thought of involves a View with buttons that change what page is being viewed. These do this with a callback function that is specified by the instantiator. This callback receives a copy of the View's state and the interaction that called the message. This should allow the scroller be very basic and unknowing of what is using it.