User Tag Scroll + Delete Button
Added a delete button to the scroller view that just deletes the message. The buttons are disabled 3 seconds before the message is deleted in on_timeout. Also may make it so that the user autocomplete is a custom trasnformer that queries for user ids which have tags, be it local or global depending on the group to display them, as opposed to the builtin discord.User transformer.