Fixed last page button
Changed the boolean which marked if the current page was the last to an int denoting the last index.
Callbacks must handle the behavior of the offset given by the scroller being much larger than the total items since the scroller has no knowledge of how many pages could exist.
Alternative solutions with callback factories may be an option so that common boilerplate is no longer needed